autom8s/projects/log_retrieval/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[project]
name = "AutoM8s"
authors = [{ name = "BSC Analytics" }]
version = "1.0.0"
requires-python = ">=3.13"
classifiers = [
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.13",
]
readme = "README.md"

dependencies = ["PyGithub~=2.6", "pandas~=2.2"]

[project.optional-dependencies]
dev = [
    "docformatter~=1.7",
    "nox~=2025.2",
    "pip-tools~=7.4",
    "pre-commit~=4.1",
    "pytest~=8.3",
    "ruff~=0.9",
    "toml~=0.10",
]
